+++ common/lanelet2_extension/include/lanelet2_extension/time/TimeConversion.h
@@ -1,0 +1,34 @@
+#pragma once
+/*
+ * Copyright (C) 2022 LEIDOS.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+#include <boost/date_time/posix_time/posix_time.hpp>
+
+namespace lanelet
+{
+
+namespace time {
+
+double toSec(const boost::posix_time::time_duration& duration);
+
+double toSec(const boost::posix_time::ptime& time);
+
+boost::posix_time::ptime timeFromSec(double sec);
+
+boost::posix_time::time_duration durationFromSec(double sec);
+
+} // namespace time
+
+}  // namespace lanelet
